// apps/console/src/__generated__/core/NewCompliancePageSubscriberDialogMutation.graphql.ts
export type MailingListSubscriberStatus = "CONFIRMED" | "PENDING";
export type CreateMailingListSubscriberInput = {
  email: string;
  fullName: string;
  mailingListId: string;
};
export type NewCompliancePageSubscriberDialogMutation$variables = {
  connections: ReadonlyArray<string>;
  input: CreateMailingListSubscriberInput;
};
export type NewCompliancePageSubscriberDialogMutation$data = {
  readonly createMailingListSubscriber: {
    readonly mailingListSubscriberEdge: {
      readonly cursor: string;
      readonly node: {
        readonly createdAt: string;
        readonly email: string;
        readonly fullName: string;
        readonly id: string;
        readonly status: MailingListSubscriberStatus;
      };
    };
  };
};
export type NewCompliancePageSubscriberDialogMutation = {
  response: NewCompliancePageSubscriberDialogMutation$data;
  variables: NewCompliancePageSubscriberDialogMutation$variables;
};